split build into parts

+++ ci/nightly.sh
@@ -2,33 +2,36 @@
 
 set -o errexit -o nounset
 
-cargo test --verbose --all
-cargo bench --verbose --all
+if [ "$KIND" == "build" ]; then
+    cargo test --verbose --all
 
-cd benches
-cargo build --all
+    BRANCH=$(if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then echo $TRAVIS_BRANCH; else echo $TRAVIS_PULL_REQUEST_BRANCH; fi)
 
-cd ../tests/run
-cargo run
+    if [ "$BRANCH" == "master" ]; then
+        echo "uploading crate docs"
 
-BRANCH=$(if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then echo $TRAVIS_BRANCH; else echo $TRAVIS_PULL_REQUEST_BRANCH; fi)
+        cargo doc --all
 
-if [ "$BRANCH" == "master" ]; then
-    echo "uploading crate docs"
+        REV=$(git rev-parse --short HEAD)
+        cd target/doc
+        rm -r .git &>/dev/null
+        git init
+        git remote add upstream "https://$GH_TOKEN@github.com/elastic-rs/elastic.git"
+        git config user.name "elastic-rs"
+        git config user.email "travis@elastic.rs"
+        git add -A .
+        git commit -qm "Build docs at ${TRAVIS_REPO_SLUG}@${REV}"
 
-    cd ../../
-    cargo doc --all
+        echo "Pushing gh-pages to GitHub"
+        git push -q upstream HEAD:refs/heads/gh-pages --force
+    fi
+elif [ "$KIND" == "bench" ]; then
+    cargo bench --verbose --all
 
-    REV=$(git rev-parse --short HEAD)
-    cd target/doc
-    rm -r .git &>/dev/null
-    git init
-    git remote add upstream "https://$GH_TOKEN@github.com/elastic-rs/elastic.git"
-    git config user.name "elastic-rs"
-    git config user.email "travis@elastic.rs"
-    git add -A .
-    git commit -qm "Build docs at ${TRAVIS_REPO_SLUG}@${REV}"
-
-    echo "Pushing gh-pages to GitHub"
-    git push -q upstream HEAD:refs/heads/gh-pages --force
+    cd benches
+    cargo build --all
+fi
+elif [ "$KIND" == "integration" ]; then 
+    cd tests/run
+    cargo run
 fi
